bs filter

+++ ocnus/src/fevm/mod.rs
@@ -28,7 +28,7 @@ use thiserror::Error;
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct FEVMData<const P: usize, FS, GS> {
     /// FEVM ensemble parameters.
-    pub params: Matrix<f64, Const<P>, Dyn, VecStorage<f64, Const<P>, Dyn>>,
+    pub params: Matrix<f32, Const<P>, Dyn, VecStorage<f32, Const<P>, Dyn>>,
 
     /// FEVM ensemble states.
     pub fevm_states: Vec<FS>,
@@ -37,7 +37,7 @@ pub struct FEVMData<const P: usize, FS, GS> {
     pub geom_states: Vec<GS>,
 
     /// Ensemble member weights.
-    pub weights: Vec<f64>,
+    pub weights: Vec<f32>,
 }
 
 impl<const P: usize, FS, GS> FEVMData<P, FS, GS>
@@ -49,10 +49,10 @@ where
     /// Create a new [`FEVMData`] filled with zeros.
     pub fn new(size: usize) -> Self {
         Self {
-            params: Matrix::<f64, Const<P>, Dyn, VecStorage<f64, Const<P>, Dyn>>::zeros(size),
+            params: Matrix::<f32, Const<P>, Dyn, VecStorage<f32, Const<P>, Dyn>>::zeros(size),
             fevm_states: vec![FS::default(); size],
             geom_states: vec![GS::default(); size],
-            weights: vec![1.0 / size as f64; size],
+            weights: vec![1.0 / size as f32; size],
         }
     }
 
@@ -76,7 +76,7 @@ where
 #[derive(Debug, Error)]
 pub enum FEVMError {
     #[error("invalid model parameter {name}={value}")]
-    InvalidParameter { name: &'static str, value: f64 },
+    InvalidParameter { name: &'static str, value: f32 },
     #[error(
         "invalid range {output_rows} x {output_cols} but expected {expected_rows} x {expected_cols}"
     )]
@@ -87,7 +87,7 @@ pub enum FEVMError {
         output_rows: usize,
     },
     #[error("attempted to simulate backwards in time (dt=-{0:.2}sec)")]
-    NegativeTimeStep(f64),
+    NegativeTimeStep(f32),
     #[error("particle filter error")]
     ParticleFilter(#[from] ParticleFilterError),
     #[error("stats error")]
@@ -108,8 +108,8 @@ where
     /// Evolve a model state forward in time.
     fn fevm_forward(
         &self,
-        time_step: f64,
-        params: &SVectorView<f64, P>,
+        time_step: f32,
+        params: &SVectorView<f32, P>,
         fevm_state: &mut FS,
         geom_state: &mut GS,
     ) -> Result<(), FEVMError>;
@@ -158,8 +158,8 @@ where
 
         debug!(
             "fevm_initialize: {:2.2}M evaluations in {:.2} sec",
-            fevmd.params.ncols() as f64 / 1e6,
-            start.elapsed().as_millis() as f64 / 1e3
+            fevmd.params.ncols() as f32 / 1e6,
+            start.elapsed().as_millis() as f32 / 1e3
         );
 
         Ok(())
@@ -203,8 +203,8 @@ where
 
         debug!(
             "fevm_initialize: {:2.2}M evaluations in {:.2} sec",
-            fevmd.params.ncols() as f64 / 1e6,
-            start.elapsed().as_millis() as f64 / 1e3
+            fevmd.params.ncols() as f32 / 1e6,
+            start.elapsed().as_millis() as f32 / 1e3
         );
 
         Ok(())
@@ -244,8 +244,8 @@ where
 
         debug!(
             "fevm_initialize_params_only: {:2.2}M evaluations in {:.2} sec",
-            fevmd.params.ncols() as f64 / 1e6,
-            start.elapsed().as_millis() as f64 / 1e3
+            fevmd.params.ncols() as f32 / 1e6,
+            start.elapsed().as_millis() as f32 / 1e3
         );
 
         Ok(())
@@ -279,8 +279,8 @@ where
 
         debug!(
             "fevm_initialize_states_only: {:2.2}M evaluations in {:.2} sec",
-            fevmd.params.ncols() as f64 / 1e6,
-            start.elapsed().as_millis() as f64 / 1e3
+            fevmd.params.ncols() as f32 / 1e6,
+            start.elapsed().as_millis() as f32 / 1e3
         );
 
         Ok(())
@@ -290,7 +290,7 @@ where
     fn fevm_observe(
         &self,
         scobs: &ScObs<ObserVec<N>>,
-        params: &SVectorView<f64, P>,
+        params: &SVectorView<f32, P>,
         fevm_state: &FS,
         geom_state: &GS,
     ) -> Result<ObserVec<N>, FEVMError>;
@@ -374,8 +374,8 @@ where
 
         debug!(
             "fevm_simulate: {:2.2}M evaluations in {:.2} sec",
-            (series.len() * fevmd.params.ncols()) as f64 / 1e6,
-            start.elapsed().as_millis() as f64 / 1e3
+            (series.len() * fevmd.params.ncols()) as f32 / 1e6,
+            start.elapsed().as_millis() as f32 / 1e3
         );
 
         let mut valid_indices_flags = vec![false; fevmd.params.ncols()];
@@ -405,7 +405,7 @@ where
     fn fevm_state(
         &self,
         series: &ScObsSeries<ObserVec<N>>,
-        params: &SVectorView<f64, P>,
+        params: &SVectorView<f32, P>,
         fevm_state: &mut FS,
         geom_state: &mut GS,
     ) -> Result<(), FEVMError>;
