+++ ocnus/src/math/bessel.rs
@@ -2,20 +2,20 @@ use crate::math::factorial;
 use std::cmp::Ordering;
 
 /// Bessel function of the first kind.
-pub fn bessel_jn(x: f64, k: usize) -> f64 {
+pub fn bessel_jn(x: f32, k: usize) -> f32 {
     match x.total_cmp(&0.0) {
-        Ordering::Equal => (matches!(k, 0) as usize) as f64,
+        Ordering::Equal => (matches!(k, 0) as usize) as f32,
         _ => {
             let sum = (1..11).try_fold((x / 2.0).powi(k as i32), |acc, idx| {
-                let next = i32::pow(-1, idx as u32) as f64 * (x / 2.0).powi((2 * idx + k) as i32)
-                    / (factorial(idx).unwrap() * factorial(idx + k).unwrap()) as f64;
+                let next = i32::pow(-1, idx as u32) as f32 * (x / 2.0).powi((2 * idx + k) as i32)
+                    / (factorial(idx).unwrap() * factorial(idx + k).unwrap()) as f32;
 
                 match next.partial_cmp(&0.0).unwrap() {
-                    std::cmp::Ordering::Greater => match next.partial_cmp(&f64::EPSILON).unwrap() {
+                    std::cmp::Ordering::Greater => match next.partial_cmp(&f32::EPSILON).unwrap() {
                         Ordering::Less => Err(acc + next),
                         _ => Ok(acc + next),
                     },
-                    std::cmp::Ordering::Less => match next.partial_cmp(&(-f64::EPSILON)).unwrap() {
+                    std::cmp::Ordering::Less => match next.partial_cmp(&(-f32::EPSILON)).unwrap() {
                         Ordering::Greater => Err(acc + next),
                         _ => Ok(acc + next),
                     },
