+++ ocnus/src/stats/mod.rs
@@ -22,13 +22,13 @@ pub enum StatsError {
     #[error("matrix does not obey required invariants: {msg}")]
     InvalidMatrix {
         msg: &'static str,
-        matrix: DMatrix<f64>,
+        matrix: DMatrix<f32>,
     },
     #[error("invalid range {name} [{minv} - {maxv}]")]
     InvalidRange {
         name: &'static str,
-        maxv: f64,
-        minv: f64,
+        maxv: f32,
+        minv: f32,
     },
 }
 /// A trait that must be implemented for any type that acts as a probability density function.
@@ -40,24 +40,24 @@ where
     ///
     /// This result is not necessarily normalized,
     /// for an exact calculation or estimate of the density see the [`PDFExactDensity`] trait.
-    fn relative_density(&self, x: &SVectorView<f64, P>) -> f64;
+    fn relative_density(&self, x: &SVectorView<f32, P>) -> f32;
 
     /// Draw a single parameter vector from the underlying density.
-    fn draw_sample(&self, rng: &mut impl Rng) -> Result<SVector<f64, P>, StatsError>;
+    fn draw_sample(&self, rng: &mut impl Rng) -> Result<SVector<f32, P>, StatsError>;
 
     /// Validate a single sample by checking for values that are out of the valid parameter range.
-    fn validate_sample(&self, sample: &SVectorView<f64, P>) -> bool {
+    fn validate_sample(&self, sample: &SVectorView<f32, P>) -> bool {
         zip_eq(sample.iter(), self.valid_range().iter()).fold(true, |acc, (c, range)| {
             acc & ((&range.0 <= c) & (c <= &range.1))
         })
     }
 
     /// Returns the valid range for parameter vector samples.
-    fn valid_range(&self) -> [(f64, f64); P];
+    fn valid_range(&self) -> [(f32, f32); P];
 }
 
 /// A trait that provides a function that calculates or estimates the exact density .
 pub trait PDFExactDensity<const P: usize>: PDF<P> {
     /// Calculates or estimates the exact density  at a specific position `x`.
-    fn exact_density(&self, x: &SVectorView<f64, P>) -> f64;
+    fn exact_density(&self, x: &SVectorView<f32, P>) -> f32;
 }
